requirements.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.5.0
autopep8==1.6.0
bcrypt==3.1.7
beautifulsoup4==4.10.0
blinker==1.4
bootstrap4==0.1.0
branca==0.4.2
Brlapi==0.8.2
certifi==2021.10.8
chardet==4.0.0
cliapp==1.20180812.1
click==7.1.2
cloudinary==1.28.0
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
confusable-homoglyphs==3.2.0
cryptography==3.3.2
cupshelpers==1.0
dbus-python==1.2.16
decorator==5.1.0
defer==1.0.6
distlib==0.3.1
distro==1.5.0
distro-info==1.0
dj-database-url==0.5.0
Django==3.2
django-bootstrap3==21.2
django-bootstrap4==21.1
django-configurations==2.3.1
django-cors-headers==3.11.0
django-daraja==1.2.0
django-google-maps==0.12.4
django-multiselectfield==0.1.12
django-on-heroku==1.1.2
django-registration==3.2
django-tinymce==3.3.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
dominate==2.6.0
duplicity==0.8.20
fasteners==0.14.1
filelock==3.0.12
Flask==1.1.4
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.0
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
Flask-WTF==1.0.0
folium==0.12.1.post1
future==0.18.2
geocoder==1.38.1
greenlet==1.1.2
gunicorn==20.1.0
gyp==0.1
httplib2==0.18.1
idna==2.10
importlib-metadata==4.0.1
itsdangerous==1.1.0
jeepney==0.7.1
Jinja2==2.11.3
keyring==23.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
lockfile==0.12.2
louis==3.18.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.4
MarkupSafe==2.0.1
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.9
numpy==1.21.5
oauthlib==3.1.0
olefile==0.46
PAM==0.4.2
paramiko==2.7.2
pexpect==4.8.0
Pillow==8.1.2
protobuf==3.12.4
psycopg2-binary==2.9.2
pycairo==1.16.2
pycodestyle==2.8.0
pycups==2.0.1
Pygments==2.7.1
PyGObject==3.40.1
PyICU==2.5
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.4.0
pyRFC3339==1.1
python-apt==2.2.1
python-dateutil==2.8.1
python-debian===0.1.39ubuntu1
python-decouple==3.5
pytz==2022.1
pyxdg==0.27
PyYAML==5.3.1
ratelim==0.1.6
reportlab==3.5.66
requests==2.25.1
SecretStorage==3.3.1
simplejson==3.17.2
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.32
sqlparse==0.4.2
systemd-python==234
toml==0.10.2
ttystatus==0.38
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.7
usb-creator==0.3.7
virtualenv==20.4.6+ds
visitor==0.1.3
wadllib==1.3.5
Werkzeug==1.0.1
whitenoise==5.3.0
WTForms==3.0.1
xdg==5
xkit==0.0.0
zipp==3.7.0
